fix: Further splitting of source code into discrete files.
