+++ lib/NativeCode/DynamicLibraryLoaderHelper/NativeRender/NativeRender.vcxproj
@@ -182,11 +182,13 @@
     <ClInclude Include="config.h" />
     <ClInclude Include="eos_minimum_includes.h" />
     <ClInclude Include="framework.h" />
+    <ClInclude Include="io_helpers.h" />
     <ClInclude Include="json.h" />
     <ClInclude Include="json_helpers.h" />
     <ClInclude Include="logging.h" />
     <ClInclude Include="pch.h" />
     <ClInclude Include="string_helpers.h" />
+    <ClInclude Include="temp.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="dllmain.cpp" />
