--- lib/NativeCode/DynamicLibraryLoaderHelper/NativeRender/json_helper.h
@@ -1,8 +1,0 @@
-#pragma once
-#include "json.h"
-
-namespace playeveryware::eos::json_helpers
-{
-    
-
-}
